Pars.Golf/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-mail==0.9.1
flask-wtf==1.2.1
wtforms==3.1.2
email-validator==2.1.0.post1
python-dotenv==1.0.1
werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
rauth==0.7.3
pyjwt==2.8.0
python-dateutil==2.8.2
pytest==7.4.4
black==24.1.1
flake8==7.0.0